Set f_globals and f_locals on _Frame
Having f_globals and f_locals in traceback frames is useful for post
mortem debugging. Thus we should provide what we can.

If we're constructing a fake traceback then we are doing so for a
Failure that has been cleaned and therefore the variables in f_globals
and f_locals are the result of _safeReprVars(). The benefit of that is
not holding references to the original objects anymore but the downside
is that the post mortem debugger only has access to string
representations of the variables. That's still a lot better than showing
all the variables as not defined in the post mortem debugger.