Make SSHSession clean up both client and session
If an environment variable passing request and a subsystem request were
both sent on the same channel, then SSHSession only cleaned up the
ISession adapter when receiving EOF or closing the channel, and did not
call loseConnection on the client transport which is the only reasonable
way for a subsystem to be notified when a connection is closed.

SSHSession now cleans up both the client transport and the ISession
adapter if both are set.