IRCClient: Allow trying message decoding with multiple codecs
The IRC protocol doesn't specify encoding, so occasionally codecs other
than "utf-8" are used. This patch allows IRCClients to specify multiple
possible codecs to use when decoding incoming messages.
Additionally, if a message can't be decoded with any of the specified
codecs, a final attempt with a non-strict Error Handler is attempted.
This defaults to "replace", but can be replaced by the user.