Send exit-status/exit-signal on EOF to SSHSession again
When I tried to upgrade https://git.launchpad.net/turnip (the hosting
backend for git.launchpad.net) to a backport of
https://github.com/twisted/twisted/pull/1696, I found that its SSH
frontend tests failed with a series of extremely opaque errors which
boiled down to git saying "fatal: the remote end hung up unexpectedly",
and which on detailed comparison were because the "exit-status" channel
request wasn't being sent after a "git upload-pack" subprocess exited.

This turns out to be because I drew an incorrect parallelism between
SSHSession.eofReceived and SSHSession.closed when fixing
https://twistedmatrix.com/trac/ticket/10308: while it is indeed
necessary to lose the client transport's connection in SSHSession.closed
even if there's a session adapter as well, it's *not* necessary to send
a close message in SSHSession.eofReceived if there's a session adapter.
Firstly, the session adapter will typically already send a close message
via SSHChannel.loseConnection once it's finished cleaning up (and would
presumably have done so prior to the fix for
https://twistedmatrix.com/trac/ticket/10308). Secondly, and more
critically, the SSH_MSG_CHANNEL_CLOSE message sent by
SSHConnection.sendClose is a request to terminate the channel, and it's
supposed to be the last thing that the party sending it sends on that
channel; SSHConnection refuses to send any other messages after sending
a close message. As a result, any asynchronous cleanup work done by
SSHSession.eofReceived, even a Deferred that fires on the next iteration
of the reactor, will fail to send cleanup messages such as "exit-status"
or "exit-signal", which the peer may interpret as abrupt termination;
that is indeed what git does.

We failed to notice this because the session tests use a StubConnection
which behaves differently: it's happy to continue to "send" messages
(i.e. record them in its various dicts of messages that would have been
sent to a real peer by the code under test) even if the channel has been
closed, and thus fails to be an accurate stub. Fixing the stub
immediately exposes the failure in
SessionInterfaceTests.test_requestExecWithData.

Reverting part of my previous change means that SSHSession correctly
sends "exit-status"/"exit-signal" messages again when a program it
started in response to a "shell" or "exec" request exits.  I've also
checked that this doesn't cause
https://twistedmatrix.com/trac/ticket/10308 to recur; only the change to
SSHSession.closed was in fact needed there.