Loosen some filenameToModule assertions
Previously the tests required filenameToModule to produce a module object with
an exactly equal `__spec__` to the one produced by an `import` statement.

However, the `import` statement varies somewhat in the result it produces and
it is unreasonable for trial to exactly match its behavior for every
point-release of Python.

Instead, change the tests so they only require that the module has the same
name and its origin seems to refer to the same path.