Grab the original object instead of expanding the proxy
The proxy is still only supposed to be to an IReporter so there's no
particular guarantee that the original object will have those attributes.
That makes the change to the proxy type a silly thing to do.  The particular
test knows exactly what the original object will be (though not such that we
can express it at the type level) so it's safe to grab the attributes there.