Factor signal handling logic out of ReactorBase (and subclasses)
ReactorBase now sets up a `SignalHandling` object with this behavior.
PosixReactorBase overrides the factory for this object to add SIGCHLD-handling
behavior.

7812431a4f03bfa62ce8d6a4fffaca1793bd7594 provides the original implementation
of the glib bits (thanks glyph).  It looks to me like we don't need to do *as
much* work as that revision does to make glib happy - but all I have to go on
is the test suite.