On further consideration, don't require a coherent state
Just deal with the incoherent state in the same way as it was dealt with in
trunk prior to this refactoring.

Some future cleanup or bugfix should fix the reactor and/or trial and/or the
Twisted test suite (and/or something else, who knows!) so that when a
`_ChildSignalHandling` is being installed, it hasn't been installed before.