Conditionally skip tests that require METHOD_CRYPT
The advent of libxcrypt offers new password hashing algorithms to linux
distributions, which we are slowly rolling out to our users.

At the same time we are removing support for legacy ciphers, that
should not be used any longer in a security-context.

One of these is the traditional crypt method, which the python
documentation¹ describes aptly as the "weakest method" it supports.

This change adds a skip condition to the relevant tests, that checks for
the presence of crypt.METHOD_CRYPT in the list of supported hashes.

[1] https://docs.python.org/3/library/crypt.html#crypt.METHOD_CRYPT

Closes: #11826