sketch of a new OpenSSL connection interface
as silly as 'IOpenSSLServerConnectionCreatorFactory' sounds as an interface
name, there is a real problem with the previous structure: Context objects are
configured by *every connection* with new ALPN/NPN parameters, which is not the
way that contexts are supposed to work.  The context configuration should
happen once per-context per-*factory*, and the connection configuration should
happen once per *connection*.  However, 'per-context' is important, because an
implementation of SNI requires that the context configuration happen on the
creation of new contexts for other hostnames that may not have been loaded yet,
not just the first context.  so it is being called both too often and not often
enough, and at the wrong time.

client-side it's still worth doing this for symmetry, although the practical
implications are somewhat less pronounced, since you just call the context
methods a bit too often.