hypothesis: some native CFRunLoop state gets GC'd between runs
the first 2 builds in this PR resulted in 5/6 macOS builds succeeding which is
seems to be slightly beating the usual odds of 50/50, and nothing has changed
except for the removal of a reference to potential active CFRunLoopSource
objects in the reactor under test.  if we *also* garbage collect in immediately
after breaking any references to CFRunLoop state, does that improve the odds
further?