Optimize Deferred._runCallbacks() to not create chain list unnecessarily
The current code effectively stores the first element of the chain in
the current variable for the duration of the loop iteration. It is
possible to optimize this code by not storing the first element in the
first place. As a result, the creation of the list and operations with
it may be avoided in many cases.