test-fix: replace test_similarZonesDontInterfere with test_domainOutsideZoneButInZoneFile
The test takes a resolver that's loaded from a zone file and makes a query to a domain name that's outside the zone, but in the zone file. This is a misconguration case. Normally, a zone file doesn't contain domains outside of the zone (as keys), except for the PTR record. For example, when bind9 loads such a zone file, it ignores the entry. The correct for the server is to respond as if it wouldn't know the record.